Move registries to utils
