+++ spyder/utils/registries.py
